+++ CotEditor/Sources/CETextView.m
@@ -1309,28 +1309,6 @@ static NSCharacterSet *kMatchingClosingBracketsSet;
 
 
 // ------------------------------------------------------
-/// trim all trailing whitespace
-- (IBAction)trimTrailingWhitespace:(nullable id)sender
-// ------------------------------------------------------
-{
-    NSMutableArray<NSString *> *replaceStrings = [NSMutableArray array];
-    NSMutableArray<NSValue *> *replaceRanges = [NSMutableArray array];
-    
-    NSRegularExpression *regex = [NSRegularExpression regularExpressionWithPattern:@"[ \\t]+$" options:NSRegularExpressionAnchorsMatchLines error:nil];
-    [regex enumerateMatchesInString:[self string] options:0
-                              range:NSMakeRange(0, [[self string] length])
-                         usingBlock:^(NSTextCheckingResult * _Nullable result, NSMatchingFlags flags, BOOL * _Nonnull stop)
-     {
-         [replaceRanges addObject:[NSValue valueWithRange:[result range]]];
-         [replaceStrings addObject:@""];
-    }];
-    
-    [self replaceWithStrings:replaceStrings ranges:replaceRanges selectedRanges:nil
-                  actionName:NSLocalizedString(@"Trim Trailing Whitespace", nil)];
-}
-
-
-// ------------------------------------------------------
 /// アウトラインメニュー選択によるテキスト選択を実行
 - (IBAction)setSelectedRangeWithNSValue:(nullable id)sender
 // ------------------------------------------------------